Reemplazar el tipo de ciertos parámetros por `size_t`

+++ EntradaSalida/Perifericos/Displays7Seg/Seriales/CLASE_Display7Segmentos1637.h
@@ -6,7 +6,7 @@
             TM1637Display *display;
             
         public:
-            Display7Segmentos1637(byte numeroInicial, bool estadoInicial, byte indiceDigito, TM1637Display *display, byte brillo);
+            Display7Segmentos1637(byte numeroInicial, bool estadoInicial, size_t indiceDigito, TM1637Display *display, byte brillo);
 
             void setNumeroBajoNivel(byte valor);
     };
